Trait jamsocket::SimpleJamsocketService

pub trait SimpleJamsocketService: Default + Unpin + Send + Sync + 'static {
    fn initialize(&mut self, room_id: &str, context: &impl JamsocketContext) { ... }

    fn connect(&mut self, user: u32, context: &impl JamsocketContext) { ... }

    fn disconnect(&mut self, user: u32, context: &impl JamsocketContext) { ... }

    fn message(
        &mut self, 
        user: u32, 
        message: &str, 
        context: &impl JamsocketContext
    ) { ... }

    fn binary(
        &mut self, 
        user: u32, 
        message: &[u8], 
        context: &impl JamsocketContext
    ) { ... }

    fn timer(&mut self, context: &impl JamsocketContext) { ... }
}

A simplified interface for creating a JamsocketService that can be exposed as a WebAssembly module.

See module documentation for usage examples.

Provided methods

fn initialize(&mut self, room_id: &str, context: &impl JamsocketContext)

Called when the service is created, before any client has had a chance to connect.

fn connect(&mut self, user: u32, context: &impl JamsocketContext)

Called each time a client connects to the service.

fn disconnect(&mut self, user: u32, context: &impl JamsocketContext)

Called each time a client disconnects from the service, unless that disconnection will cause the service to be destroyed.

fn message(&mut self, user: u32, message: &str, context: &impl JamsocketContext)

Called each time a client sends a text message to the service.

fn binary(&mut self, user: u32, message: &[u8], context: &impl JamsocketContext)

Called each time a client sends a binary message to the service.

fn timer(&mut self, context: &impl JamsocketContext)

Called when JamsocketContext::set_timer has been called on this service’s context, after the provided duration.